linux: add 2.6.29 kernel for h3600
Add a recipe for Linux 2.6.29 for the iPAQ h3600 machine only. (Note
that this will probably not work for h3800 as broken support for this
machine is removed by one of the patches.)

Kernel defconfig and patches by Dmitry Artamonow <mad_soft@inbox.ru>
(downloaded from http://mad.is-a-geek.org/h3600/).

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>