linux-yocto-tiny-kexecboot: initial defconfig updates for 3.10
Zaurus configuration was based on the 3.2 kernel and when the
recipe was removed we simply upgraded whithout much testing.

This first pass of cleaning should help to standardize and
possibly move back using the defconfig made with do_savedefconfig.

Those are much shorter and generally don't change much.
Tested on poodle,c7x0 and collie (CF only).

Signed-off-by: Andrea Adami <andrea.adami@gmail.com>