linux-yocto*: collie: update Partition Configuration Register patch
The CFI readings of the number of hardware partitions is now fixed
and correctly showing 2.
Further output is valid only for the default case of 2 asymmetric partitions
but the CFI code in kernel does only care about the number of equal partitions.

So we set the PCR accordingly but with 2 symmetric partitions of 2 planes each
without loosing the advantages of the Dual Work.

Signed-off-by: Andrea Adami <andrea.adami@gmail.com>