gcw0: improve compatibility with kexecboot
Alternative images to be booted by kexecboot
require to have an elf vmlinuz in /boot (kexec
on mips expects elf).
The bootloader however needs raw data (vmlinuz.bin) so
we customize KERNEL_IMAGETYPE in the recipe.
Finally, kexec is added to the main kernel recipe for
easier testing and the watchdog timer is increased in
the configuration for the kexecboot kernel.

At the moment linux-kexecboot detects just fine
partitions on external SD but freezes on kexec.
Tests done from shell reveal that both kexec
binaries (2.06 glibc and 2.02 klibc) load the new kernel
but freeze on kexec -e.

Signed-off-by: Andrea Adami <andrea.adami@gmail.com>